requirements.txt
attrs==22.1.0
black==22.10.0
click==8.1.3
exceptiongroup==1.0.4
Flask==2.2.2
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
psycopg2==2.9.5
pyparsing==3.0.9
pytest==7.2.0
pytest-mock==3.10.0
python-dotenv==0.21.0
tomli==2.0.1
Werkzeug==2.2.2
